Move container-feeder to an optional test.
Container-feeder is only needed, when the node is not running against a
registry. This should only be the case for SLE12.

If the node does not use registry-image (use as default case here) the
container-feeder is expected to be running.

(cherry picked from commit 91408dd29a858a2530f67de192de7688791f24d6)