Jenkinsfile: Use docker cmdline directly instead of k8s Jenkins plugin
The codenarc and jenkins-file pipelines are the only ones which depends on the k8s
Jenkins plugin. As such, we can use docker directly in order to be able to
drop the plugin from the server. The nodelabel is hardcoded because it
does not make much sense to make this configurable given everything happens
on a container.