misc-tools: upload-image-openstack: Fix conditional for glance images
The openstack command always returns 0 if the command was successful
even when an image was not present. As such, we need to actually check
the output of this command to find out if we already have the image on
the server.

(cherry picked from commit af503d9391a521b1d9fd5def0294b0b7cd951bcf)